refactor(spectral tools): AColor and dependencies moved to spectrogram module
