+++ src/spectrogram/AColor.h
