+++ src/spectrogram/CMakeLists.txt
@@ -5,13 +5,21 @@
 declare_module(spectrogram)
 
 set(MODULE_SRC
+    ${CMAKE_CURRENT_LIST_DIR}/AColor.h
     ${CMAKE_CURRENT_LIST_DIR}/ispectrogramconfiguration.h
     ${CMAKE_CURRENT_LIST_DIR}/spectrogrammodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/spectrogrammodule.h
     ${CMAKE_CURRENT_LIST_DIR}/spectrogramtypes.h
 
+    ${CMAKE_CURRENT_LIST_DIR}/internal/AColor.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/AColorResources.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/AllThemeResources.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/AllThemeResources.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/MacroMagic.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramconfiguration.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/spectrogramconfiguration.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/Theme.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/Theme.h
 
     ${CMAKE_CURRENT_LIST_DIR}/view/abstractsectionparameterslistmodel.cpp
     ${CMAKE_CURRENT_LIST_DIR}/view/abstractsectionparameterslistmodel.h
