+++ src/spectrogram/internal/AColorResources.h
