+++ src/spectrogram/internal/AllThemeResources.cpp
