+++ src/spectrogram/internal/AllThemeResources.h
