+++ src/spectrogram/internal/MacroMagic.h
