+++ src/spectrogram/internal/Theme.cpp
